# Conflicts:
#	.idea/modules.xml

+++ finmartserviceapi/src/main/java/magicfinmart/datacomp/com/finmartserviceapi/motor/model/AddonEntity.java
@@ -25,193 +25,193 @@ public class AddonEntity {
      * addon_final_premium : 0
      */
 
-    private int addon_zero_dep_cover;
-    private int addon_road_assist_cover;
-    private int addon_ncb_protection_cover;
-    private int addon_engine_protector_cover;
-    private int addon_invoice_price_cover;
-    private int addon_key_lock_cover;
-    private int addon_consumable_cover;
-    private int addon_daily_allowance_cover;
-    private int addon_windshield_cover;
-    private int addon_passenger_assistance_cover;
-    private int addon_tyre_coverage_cover;
-    private int addon_personal_belonging_loss_cover;
-    private int addon_inconvenience_allowance_cover;
-    private int addon_medical_expense_cover;
-    private int addon_hospital_cash_cover;
-    private int addon_ambulance_charge_cover;
-    private int addon_rodent_bite_cover;
-    private int addon_losstime_protection_cover;
-    private int addon_hydrostatic_lock_cover;
-    private int addon_guaranteed_auto_protection_cover;
-    private int addon_final_premium;
+    private double addon_zero_dep_cover;
+    private double addon_road_assist_cover;
+    private double addon_ncb_protection_cover;
+    private double addon_engine_protector_cover;
+    private double addon_invoice_price_cover;
+    private double addon_key_lock_cover;
+    private double addon_consumable_cover;
+    private double addon_daily_allowance_cover;
+    private double addon_windshield_cover;
+    private double addon_passenger_assistance_cover;
+    private double addon_tyre_coverage_cover;
+    private double addon_personal_belonging_loss_cover;
+    private double addon_inconvenience_allowance_cover;
+    private double addon_medical_expense_cover;
+    private double addon_hospital_cash_cover;
+    private double addon_ambulance_charge_cover;
+    private double addon_rodent_bite_cover;
+    private double addon_losstime_protection_cover;
+    private double addon_hydrostatic_lock_cover;
+    private double addon_guaranteed_auto_protection_cover;
+    private double addon_final_premium;
 
-    public int getAddon_zero_dep_cover() {
+    public double getAddon_zero_dep_cover() {
         return addon_zero_dep_cover;
     }
 
-    public void setAddon_zero_dep_cover(int addon_zero_dep_cover) {
+    public void setAddon_zero_dep_cover(double addon_zero_dep_cover) {
         this.addon_zero_dep_cover = addon_zero_dep_cover;
     }
 
-    public int getAddon_road_assist_cover() {
+    public double getAddon_road_assist_cover() {
         return addon_road_assist_cover;
     }
 
-    public void setAddon_road_assist_cover(int addon_road_assist_cover) {
+    public void setAddon_road_assist_cover(double addon_road_assist_cover) {
         this.addon_road_assist_cover = addon_road_assist_cover;
     }
 
-    public int getAddon_ncb_protection_cover() {
+    public double getAddon_ncb_protection_cover() {
         return addon_ncb_protection_cover;
     }
 
-    public void setAddon_ncb_protection_cover(int addon_ncb_protection_cover) {
+    public void setAddon_ncb_protection_cover(double addon_ncb_protection_cover) {
         this.addon_ncb_protection_cover = addon_ncb_protection_cover;
     }
 
-    public int getAddon_engine_protector_cover() {
+    public double getAddon_engine_protector_cover() {
         return addon_engine_protector_cover;
     }
 
-    public void setAddon_engine_protector_cover(int addon_engine_protector_cover) {
+    public void setAddon_engine_protector_cover(double addon_engine_protector_cover) {
         this.addon_engine_protector_cover = addon_engine_protector_cover;
     }
 
-    public int getAddon_invoice_price_cover() {
+    public double getAddon_invoice_price_cover() {
         return addon_invoice_price_cover;
     }
 
-    public void setAddon_invoice_price_cover(int addon_invoice_price_cover) {
+    public void setAddon_invoice_price_cover(double addon_invoice_price_cover) {
         this.addon_invoice_price_cover = addon_invoice_price_cover;
     }
 
-    public int getAddon_key_lock_cover() {
+    public double getAddon_key_lock_cover() {
         return addon_key_lock_cover;
     }
 
-    public void setAddon_key_lock_cover(int addon_key_lock_cover) {
+    public void setAddon_key_lock_cover(double addon_key_lock_cover) {
         this.addon_key_lock_cover = addon_key_lock_cover;
     }
 
-    public int getAddon_consumable_cover() {
+    public double getAddon_consumable_cover() {
         return addon_consumable_cover;
     }
 
-    public void setAddon_consumable_cover(int addon_consumable_cover) {
+    public void setAddon_consumable_cover(double addon_consumable_cover) {
         this.addon_consumable_cover = addon_consumable_cover;
     }
 
-    public int getAddon_daily_allowance_cover() {
+    public double getAddon_daily_allowance_cover() {
         return addon_daily_allowance_cover;
     }
 
-    public void setAddon_daily_allowance_cover(int addon_daily_allowance_cover) {
+    public void setAddon_daily_allowance_cover(double addon_daily_allowance_cover) {
         this.addon_daily_allowance_cover = addon_daily_allowance_cover;
     }
 
-    public int getAddon_windshield_cover() {
+    public double getAddon_windshield_cover() {
         return addon_windshield_cover;
     }
 
-    public void setAddon_windshield_cover(int addon_windshield_cover) {
+    public void setAddon_windshield_cover(double addon_windshield_cover) {
         this.addon_windshield_cover = addon_windshield_cover;
     }
 
-    public int getAddon_passenger_assistance_cover() {
+    public double getAddon_passenger_assistance_cover() {
         return addon_passenger_assistance_cover;
     }
 
-    public void setAddon_passenger_assistance_cover(int addon_passenger_assistance_cover) {
+    public void setAddon_passenger_assistance_cover(double addon_passenger_assistance_cover) {
         this.addon_passenger_assistance_cover = addon_passenger_assistance_cover;
     }
 
-    public int getAddon_tyre_coverage_cover() {
+    public double getAddon_tyre_coverage_cover() {
         return addon_tyre_coverage_cover;
     }
 
-    public void setAddon_tyre_coverage_cover(int addon_tyre_coverage_cover) {
+    public void setAddon_tyre_coverage_cover(double addon_tyre_coverage_cover) {
         this.addon_tyre_coverage_cover = addon_tyre_coverage_cover;
     }
 
-    public int getAddon_personal_belonging_loss_cover() {
+    public double getAddon_personal_belonging_loss_cover() {
         return addon_personal_belonging_loss_cover;
     }
 
-    public void setAddon_personal_belonging_loss_cover(int addon_personal_belonging_loss_cover) {
+    public void setAddon_personal_belonging_loss_cover(double addon_personal_belonging_loss_cover) {
         this.addon_personal_belonging_loss_cover = addon_personal_belonging_loss_cover;
     }
 
-    public int getAddon_inconvenience_allowance_cover() {
+    public double getAddon_inconvenience_allowance_cover() {
         return addon_inconvenience_allowance_cover;
     }
 
-    public void setAddon_inconvenience_allowance_cover(int addon_inconvenience_allowance_cover) {
+    public void setAddon_inconvenience_allowance_cover(double addon_inconvenience_allowance_cover) {
         this.addon_inconvenience_allowance_cover = addon_inconvenience_allowance_cover;
     }
 
-    public int getAddon_medical_expense_cover() {
+    public double getAddon_medical_expense_cover() {
         return addon_medical_expense_cover;
     }
 
-    public void setAddon_medical_expense_cover(int addon_medical_expense_cover) {
+    public void setAddon_medical_expense_cover(double addon_medical_expense_cover) {
         this.addon_medical_expense_cover = addon_medical_expense_cover;
     }
 
-    public int getAddon_hospital_cash_cover() {
+    public double getAddon_hospital_cash_cover() {
         return addon_hospital_cash_cover;
     }
 
-    public void setAddon_hospital_cash_cover(int addon_hospital_cash_cover) {
+    public void setAddon_hospital_cash_cover(double addon_hospital_cash_cover) {
         this.addon_hospital_cash_cover = addon_hospital_cash_cover;
     }
 
-    public int getAddon_ambulance_charge_cover() {
+    public double getAddon_ambulance_charge_cover() {
         return addon_ambulance_charge_cover;
     }
 
-    public void setAddon_ambulance_charge_cover(int addon_ambulance_charge_cover) {
+    public void setAddon_ambulance_charge_cover(double addon_ambulance_charge_cover) {
         this.addon_ambulance_charge_cover = addon_ambulance_charge_cover;
     }
 
-    public int getAddon_rodent_bite_cover() {
+    public double getAddon_rodent_bite_cover() {
         return addon_rodent_bite_cover;
     }
 
-    public void setAddon_rodent_bite_cover(int addon_rodent_bite_cover) {
+    public void setAddon_rodent_bite_cover(double addon_rodent_bite_cover) {
         this.addon_rodent_bite_cover = addon_rodent_bite_cover;
     }
 
-    public int getAddon_losstime_protection_cover() {
+    public double getAddon_losstime_protection_cover() {
         return addon_losstime_protection_cover;
     }
 
-    public void setAddon_losstime_protection_cover(int addon_losstime_protection_cover) {
+    public void setAddon_losstime_protection_cover(double addon_losstime_protection_cover) {
         this.addon_losstime_protection_cover = addon_losstime_protection_cover;
     }
 
-    public int getAddon_hydrostatic_lock_cover() {
+    public double getAddon_hydrostatic_lock_cover() {
         return addon_hydrostatic_lock_cover;
     }
 
-    public void setAddon_hydrostatic_lock_cover(int addon_hydrostatic_lock_cover) {
+    public void setAddon_hydrostatic_lock_cover(double addon_hydrostatic_lock_cover) {
         this.addon_hydrostatic_lock_cover = addon_hydrostatic_lock_cover;
     }
 
-    public int getAddon_guaranteed_auto_protection_cover() {
+    public double getAddon_guaranteed_auto_protection_cover() {
         return addon_guaranteed_auto_protection_cover;
     }
 
-    public void setAddon_guaranteed_auto_protection_cover(int addon_guaranteed_auto_protection_cover) {
+    public void setAddon_guaranteed_auto_protection_cover(double addon_guaranteed_auto_protection_cover) {
         this.addon_guaranteed_auto_protection_cover = addon_guaranteed_auto_protection_cover;
     }
 
-    public int getAddon_final_premium() {
+    public double getAddon_final_premium() {
         return addon_final_premium;
     }
 
-    public void setAddon_final_premium(int addon_final_premium) {
+    public void setAddon_final_premium(double addon_final_premium) {
         this.addon_final_premium = addon_final_premium;
     }
 }
